# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Teensy
  description: Tiny Url Generator
  version: 1.0.0
  contact:
    name: Tejas Ozarkar
    email: tejozarkar@gmail.com
servers:
  - url: http://localhost:8080
    description: Local Environment
  - url: http://teensyapi-env.eba-k2xfeapv.ap-south-1.elasticbeanstalk.com
    description: Production Environment
paths:
  /approval:
    post:
      tags:
        - group
      summary: Send approval to group admins
      operationId: sendApproval
      requestBody:
        description: Send approval to admin
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDto'
      responses:
        200:
          description: Approval Message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
...
  /auth/signup:
    post:
      summary:
      operationId: signup
      requestBody:
        description: Create new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        200:
          description: User Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
  /auth/user:
    get:
      summary: Get User Info
      operationId: getUserInfo
      responses:
        200:
          description: Return User Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        401:
          description: Unauthorized User
  /auth/signin:
    post:
      summary:
      operationId: signin
      requestBody:
        description: Logging in user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        200:
          description: User Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
components:
  schemas:
    ApprovalDto:
      allOf:
        - $ref: '#/components/schemas/CardDto'
      type: object
      properties:
        groupId:
          type: int
        cardId:
          type: int
        longurl:
          type: string
...
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        jwt:
          type: string
        username:
          type: string
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
    UserDto:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        username:
          type: string
        password:
          type: string
        email:
          type: string
...
    CardDto:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
...
    GenericResponse:
      type: object
      properties:
        message:
          type: string